big refactor and add web based validation
