+++ tsconfig.json
@@ -7,8 +7,11 @@
         "src/proofPointEvent.ts",
         "src/proofPointEventType.ts",
         "src/proofPointIssueResult.ts",
-        "src/proofPointRegistry.ts",
-        "src/proofPointRegistryRoot.ts",
+        "src/ethereumProofPointRegistry.ts",
+        "src/ethereumProofPointRegistryRoot.ts",
+        "src/proofPointValidator.ts",
+        "src/proofPointResolver.ts",
+        "src/proofPointAuthenticator.ts",
         "src/proofPointStatus.ts",
         "src/canonicalize.d.ts",
         "src/local-iso-dt.d.ts",
